(PC-26911) refactor(login): Handle signin success in query file
